[create topic flow] rebase
