+++ src/components/forums/forum/ForumContent.tsx
@@ -1,10 +1,9 @@
 import isNil from 'lodash/isNil';
 import { PublicKey } from '@solana/web3.js';
 import Markdown from 'markdown-to-jsx';
-import { useState, ReactNode, useEffect } from 'react';
+import { useState, ReactNode, useEffect, useMemo, useCallback } from 'react';
 import ReactGA from 'react-ga4';
 import Lottie from 'lottie-react';
-import { PostRestriction } from '@usedispatch/client';
 
 import { Lock, Plus, Trash } from '../../../assets';
 import animationData from '../../../lotties/loader.json';
@@ -25,12 +24,12 @@ import {
   UploadForumBanner,
   ConnectionAlert,
   Notification,
+  CreateTopic,
 } from '..';
 import { StarsAlert } from '../StarsAlert';
-import { useRole } from '../../../contexts/DispatchProvider';
 
 import { DispatchForum } from '../../../utils/postbox/postboxWrapper';
-import { SCOPES, UserRoleType } from '../../../utils/permissions';
+import { SCOPES } from '../../../utils/permissions';
 import { DispatchError, Result } from '../../../types/error';
 import { isError, errorSummary } from '../../../utils/error';
 import { isSuccess } from '../../../utils/loading';
@@ -46,6 +45,8 @@ import {
 } from '../../../utils/restrictionListHelper';
 import { newPublicKey } from '../../../utils/postbox/validateNewPublicKey';
 import { csvStringToPubkeyList } from '../../../utils/csvStringToPubkeyList';
+import { ForumPost } from '@usedispatch/client';
+import { usePath } from '../../../contexts/DispatchProvider';
 
 interface ForumContentProps {
   forumObject: DispatchForum;
@@ -109,8 +110,7 @@ export function ForumContent(props: ForumContentProps): JSX.Element {
           </div>
           <div className="toolsWrapper" />
           <div className="topicListWrapper">
-            <TopicList
-             />
+            <TopicList />
           </div>
         </div>
       </div>
@@ -138,8 +138,8 @@ interface PopulatedForumContentProps {
 
 export function PopulatedForumContent(props: PopulatedForumContentProps): JSX.Element {
   const { initialForumData, forumObject, update } = props;
-  const { roles } = useRole();
   const { permission } = forumObject;
+  const { buildTopicPath } = usePath();
 
   const [forumData, setForumData] = useState<ForumData>(initialForumData);
 
@@ -150,23 +150,8 @@ export function PopulatedForumContent(props: PopulatedForumContentProps): JSX.El
     setForumData({ ...forumData, images: { ...images, background: imageUrl } });
     await update();
   };
-  const [newTopic, setNewTopic] = useState<{
-    title: string;
-    description: string;
-    NFTaccessToken: string;
-    SPLaccessToken: string;
-    SPLamount: number;
-  }>({
-    title: '',
-    description: '',
-    NFTaccessToken: '',
-    SPLaccessToken: '',
-    SPLamount: 1,
-  });
 
-  const [showNewTopicModal, setShowNewTopicModal] = useState(false);
-  const [creatingNewTopic, setCreatingNewTopic] = useState(false);
-  const [newTopicInFlight, setNewTopicInFlight] = useState(false);
+  const [newTopicInFlight, setNewTopicInFlight] = useState<{ title: string }>();
   const [addNFTGate, setAddNFTGate] = useState(false);
   const [addSPLGate, setAddSPLGate] = useState(false);
   const [tokenGateSelection, setTokenGateSelection] = useState('');
@@ -190,8 +175,7 @@ export function PopulatedForumContent(props: PopulatedForumContentProps): JSX.El
       return restrictionListToString(forumData.restriction);
     } else return [];
   });
-  const [ungatedNewTopic, setUngatedNewTopic] = useState(currentForumAccessToken.length === 0);
-  const [keepGates, setKeepGates] = useState(!ungatedNewTopic);
+
   const [newForumAccessToken, setNewForumAccessToken] = useState<string>('');
   const [newForumAccessTokenAmount, setNewForumAccessTokenAmount] =
     useState<number>(1);
@@ -346,132 +330,6 @@ export function PopulatedForumContent(props: PopulatedForumContentProps): JSX.El
     }
   };
 
-  const createTopic = async (): Promise<void> => {
-    const p = {
-      subj: newTopic.title,
-      body: newTopic.description,
-    };
-
-    setCreatingNewTopic(true);
-    let restrictionResult: Result<PostRestriction> | undefined;
-    // First case checks if existing gates are kept and new ones being added
-    // Second case removes existing gates and adds new ones
-    // Third case removes existing gates
-    // Final case keeps existing gates
-    if (keepGates && addNFTGate && newTopic.NFTaccessToken !== '') {
-      restrictionResult = pubkeysToRestriction(
-        newTopic.NFTaccessToken,
-        isSuccess(forumData.restriction) ? forumData.restriction : undefined,
-      );
-    } else if (!keepGates && addNFTGate && newTopic.NFTaccessToken !== '') {
-      restrictionResult = pubkeysToRestriction(newTopic.NFTaccessToken);
-    } else if (
-      (!isSuccess(forumData.restriction) || ungatedNewTopic) &&
-      addSPLGate &&
-      newTopic.SPLaccessToken !== ''
-    ) {
-      // TODO: turn into a util function later
-      try {
-        const splMint = newPublicKey(newTopic.SPLaccessToken);
-        if (isSuccess(splMint)) {
-          const tokenMetadata = await forumObject.connection.getTokenSupply(
-            splMint,
-          );
-          restrictionResult = pubkeysToSPLRestriction(
-            newTopic.SPLaccessToken,
-            newTopic.SPLamount,
-            tokenMetadata.value.decimals,
-          );
-        }
-      } catch (error) {
-        setCreatingNewTopic(false);
-        setModalInfo({
-          title: 'Something went wrong!',
-          type: MessageType.error,
-          body: 'The topic could not be created',
-          collapsible: { header: 'Error', content: errorSummary(error as DispatchError) },
-        });
-        setShowNewTopicModal(false);
-        return;
-      }
-    } else if (!keepGates) {
-      restrictionResult = { null: {} };
-    } else {
-      // No restriction
-      restrictionResult = undefined;
-    }
-    if (isError(restrictionResult)) {
-      const error = restrictionResult;
-      setCreatingNewTopic(false);
-      setModalInfo({
-        title: 'Something went wrong!',
-        type: MessageType.error,
-        body: 'The topic could not be created',
-        collapsible: { header: 'Error', content: errorSummary(error) },
-      });
-      setShowNewTopicModal(false);
-      return;
-    }
-
-    // the possibility of error is no longer present, so rename
-    // this variable restriction
-    const restriction = restrictionResult;
-
-    const tx = await forumObject.createTopic(
-      p,
-      forumData.collectionId,
-      restriction,
-    );
-    if (isSuccess(tx)) {
-      setCreatingNewTopic(false);
-      setNewTopicInFlight(true);
-      setModalInfo({
-        body: <TransactionLink transaction={tx} />,
-        type: MessageType.success,
-        title: 'Topic created!',
-      });
-      setNewTopic({
-        title: '',
-        description: '',
-        NFTaccessToken: '',
-        SPLaccessToken: '',
-        SPLamount: 1,
-      });
-      setShowNewTopicModal(false);
-
-      // re-load forum in background
-      await forumObject.connection.confirmTransaction(tx).then(() => {
-        void update();
-        setNewTopicInFlight(false);
-      });
-    } else {
-      const error = tx;
-      setCreatingNewTopic(false);
-      setModalInfo({
-        title: 'Something went wrong!',
-        type: MessageType.error,
-        body: 'The topic could not be created',
-        collapsible: { header: 'Error', content: errorSummary(error) },
-      });
-      setShowNewTopicModal(false);
-    }
-  };
-
-  const createTopicButton = (
-    <button
-      className={'createTopicButton'}
-      type="button"
-      disabled={!permission.readAndWrite}
-      onClick={() => {
-        setShowNewTopicModal(true);
-      }}>
-      <div className="buttonImageContainer">
-        <Plus />
-      </div>
-      Create Topic
-    </button>
-  );
-
   const parseCollectionList = (accessListString): void => {
     try {
       if (accessListString.length > 0) {
@@ -490,7 +348,17 @@ export function PopulatedForumContent(props: PopulatedForumContentProps): JSX.El
     }
   };
 
-  const forumHeader = (
+  const onTopicCreated = useCallback((): void => {
+    console.log('oncreate', initialForumData.posts);
+    const topics = initialForumData.posts.filter(p => p.isTopic);
+    console.log('selected', topics, topics[0]);
+    const topicPath = buildTopicPath(initialForumData.collectionId.toBase58(), (topics[0] as ForumPost).postId);
+    console.log('path', topicPath);
+    setNewTopicInFlight(undefined);
+    // location.assign(`${forumPath}${location.search}`);
+  }, [initialForumData]);
+
+  const forumHeader = useMemo(() => (
     <div className="forumContentHeader">
       <div className={'titleBox'}>
         {currentForumAccessToken.length > 0 && (
@@ -505,25 +373,23 @@ export function PopulatedForumContent(props: PopulatedForumContentProps): JSX.El
         <div className="description">
           <Markdown>{forumData.description.desc}</Markdown>
         </div>
-        {createTopicButton}
+        <CreateTopic
+          forumObject={forumObject}
+          forumData={initialForumData}
+          currentForumAccessToken={currentForumAccessToken}
+          topicInFlight={(title) => {
+            if (title === '') {
+              setNewTopicInFlight(undefined);
+            } else {
+              setNewTopicInFlight({ title });
+            }
+          }}
+          update={update}
+          onTopicCreated={onTopicCreated}
+        />
       </div>
     </div>
-  );
-
-  useEffect(() => {
-    if (
-      !keepGates &&
-      (newTopic.NFTaccessToken.length === 0 ||
-        newTopic.SPLaccessToken.length === 0)
-    ) {
-      setUngatedNewTopic(true);
-    } else {
-      setUngatedNewTopic(false);
-      if (tokenGateSelection === 'SPL') {
-        setTokenGateSelection('');
-      }
-    }
-  }, [newTopic.SPLaccessToken, newTopic.NFTaccessToken.length, keepGates]);
+  ), [initialForumData]);
 
   useEffect(() => {
     parseCollectionList(newForumAccessToken);
@@ -722,192 +588,6 @@ export function PopulatedForumContent(props: PopulatedForumContentProps): JSX.El
               }
             />
           )}
-          {(() => {
-            if (showNewTopicModal && isNil(modalInfo)) {
-              if (roles.includes(UserRoleType.Viewer)) {
-                return (
-                  <PopUpModal
-                    id="create-topic"
-                    title="You are not authorized"
-                    body={
-                      isSuccess(forumData.restriction) &&
-                      !isNil(forumData.restriction.tokenOwnership) &&
-                      isSuccess(forumData.restriction.tokenOwnership) &&
-                      forumData.restriction.tokenOwnership.mint.equals(
-                        forumData.moderatorMint,
-                      )
-                        ? 'Oops! Only moderators can create new topics at this time.'
-                        : 'Oops! You need a token to participate. Please contact the forum’s moderators.'
-                    }
-                    visible
-                    okButton={
-                      <button
-                        className="okButton"
-                        onClick={() => setShowNewTopicModal(false)}>
-                        OK
-                      </button>
-                    }
-                    onClose={() => setShowNewTopicModal(false)}
-                  />
-                );
-              } else {
-                return (
-                  <PopUpModal
-                    id="create-topic"
-                    visible
-                    title={'Create new Topic'}
-                    onClose={() => setShowNewTopicModal(false)}
-                    body={
-                      <div className="createTopicBody">
-                        <>
-                          <span className="createTopicLabel">Topic title</span>
-                          <input
-                            type="text"
-                            placeholder="Title"
-                            className="createTopicTitleInput"
-                            name="name"
-                            required
-                            value={newTopic.title}
-                            onChange={e =>
-                              setNewTopic({
-                                ...newTopic,
-                                title: e.target.value,
-                              })
-                            }
-                          />
-                        </>
-                        <>
-                          <span className="createTopicLabel">
-                            Topic description
-                          </span>
-                          <textarea
-                            placeholder="Description"
-                            className="createTopicTitleInput createTopicTextArea"
-                            maxLength={800}
-                            value={newTopic.description}
-                            onChange={e =>
-                              setNewTopic({
-                                ...newTopic,
-                                description: e.target.value,
-                              })
-                            }
-                          />
-                        </>
-                        <PermissionsGate
-                          scopes={[SCOPES.canAddTopicRestriction]}>
-                          <>
-                            {currentForumAccessToken.length > 0 && (
-                              <div className="gateCheckbox">
-                                <div className="createTopicLabel">
-                                  Keep existing forum gates on topic
-                                </div>
-                                <input
-                                  type="checkbox"
-                                  checked={keepGates}
-                                  onChange={e => {
-                                    setKeepGates(e.target.checked);
-                                  }}
-                                />
-                              </div>
-                            )}
-                            <div className="gateCheckbox">
-                              <div className="createTopicLabel">
-                                Add a Token Gate?
-                              </div>
-                              <select
-                                value={tokenGateSelection}
-                                className="addTokenGateSelect"
-                                onChange={e =>
-                                  setTokenGateSelection(e.target.value)
-                                }>
-                                <option value="">Select a token type</option>
-                                <option value="NFT">Metaplex NFT</option>
-                                {ungatedNewTopic && (
-                                  <option value="SPL">SPL Token</option>
-                                )}
-                              </select>
-                            </div>
-                            {addNFTGate && (
-                              <div>
-                                <span className="createTopicLabel">
-                                  Limit post access by NFT Collection ID
-                                </span>
-                                <input
-                                  type="text"
-                                  placeholder="Metaplex collection ID as comma separated list"
-                                  className="newAccessToken"
-                                  name="accessToken"
-                                  value={newTopic.NFTaccessToken}
-                                  onChange={e =>
-                                    setNewTopic({
-                                      ...newTopic,
-                                      NFTaccessToken: e.target.value,
-                                    })
-                                  }
-                                />
-                              </div>
-                            )}
-                            {addSPLGate && (
-                              <div className="addSPLToken">
-                                <span className="createTopicLabel">
-                                  Limit post access by SPL Mint ID
-                                </span>
-                                <input
-                                  type="text"
-                                  placeholder="Token mint ID"
-                                  className="newAccessToken"
-                                  name="accessToken"
-                                  value={newTopic.SPLaccessToken}
-                                  onChange={e => {
-                                    setNewTopic({
-                                      ...newTopic,
-                                      SPLaccessToken: e.target.value,
-                                    });
-                                  }}
-                                />
-                                <span className="createTopicLabel">Amount</span>
-                                <Input
-                                  type="number"
-                                  value={1.0}
-                                  min={0}
-                                  step={0.01}
-                                  className="newAccessToken"
-                                  onChange={value => {
-                                    setNewTopic({
-                                      ...newTopic,
-                                      SPLamount: parseFloat(value),
-                                    });
-                                  }}
-                                />
-                              </div>
-                            )}
-                          </>
-                        </PermissionsGate>
-                      </div>
-                    }
-                    loading={creatingNewTopic}
-                    okButton={
-                      <button
-                        className="okButton"
-                        disabled={newTopic.title.length === 0}
-                        onClick={async () => createTopic()}>
-                        Create
-                      </button>
-                    }
-                    cancelButton={
-                      <button
-                        className="cancelButton"
-                        onClick={() => setShowNewTopicModal(false)}>
-                        Cancel
-                      </button>
-                    }
-                  />
-                );
-              }
-            } else {
-              return null;
-            }
-          })()}
           <div
             className="forumContentBox"
             style={{
@@ -955,17 +635,11 @@ export function PopulatedForumContent(props: PopulatedForumContentProps): JSX.El
               </div>
             </PermissionsGate>
           </div>
-          {(() => {
-            if (newTopicInFlight) {
-              return <Spinner />;
-            } else if (!isNil(forumData.collectionId)) {
-              return (
-                <div className="topicListWrapper">
-                  <TopicList forumData={forumData} />
-                </div>
-              );
-            }
-          })()}
+          {!isNil(forumData.collectionId) && (
+            <div className="topicListWrapper">
+              <TopicList forumData={forumData} topicInFlight={newTopicInFlight}/>
+            </div>
+          )}
         </>
       </div>
     </div>
