+++ src/components/forums/forum/TopicList.tsx
@@ -1,6 +1,8 @@
 import { isNil } from 'lodash';
 import { useMemo } from 'react';
 
+import { TopicInFlightRow, TopicListRow } from './TopicListRow';
+
 import {
   selectTopics,
   sortByVotes,
@@ -8,13 +10,12 @@ import {
 } from '../../../utils/posts';
 import { ForumData } from '../../../utils/hooks';
 
-import { TopicListRow } from './TopicListRow';
-
 interface TopicListProps {
   forumData?: ForumData;
+  topicInFlight?: { title: string };
 }
 
-export function TopicList({ forumData }: TopicListProps): JSX.Element {
+export function TopicList({ forumData, topicInFlight }: TopicListProps): JSX.Element {
   if (isNil(forumData)) {
     return (
     <div className="topicListContainer">
@@ -71,6 +72,9 @@ export function TopicList({ forumData }: TopicListProps): JSX.Element {
             </tr>
           </thead>
           <tbody>
+            {!isNil(topicInFlight) && (
+              <TopicInFlightRow title={topicInFlight.title} />
+            )}
             {topics.map((topic, index) => (
               <TopicListRow key={index} topic={topic} forumData={forumData} />
             ))}
