+++ src/components/forums/index.tsx
@@ -3,6 +3,7 @@ export * from './Notification';
 export * from './PoweredByDispatch';
 
 export * from './forum/CreateForum';
+export * from './forum/CreateTopic';
 export * from './forum/EditForum';
 export * from './forum/ForumContent';
 export * from './forum/ManageModerators';
