+++ src/style.css
@@ -18,6 +18,7 @@
 
 /* Forum */
 @import './styles/CreateForum.css';
+@import './styles/CreateTopic.css';
 @import './styles/EditForum.css';
 @import './styles/ForumContent.css';
 @import './styles/ManageOwners.css';
