+++ src/styles/CreateTopic.css
@@ -1,0 +1,181 @@
+@layer components {
+  .createTopicButton {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    white-space: nowrap;
+
+    padding: 16px;
+    border-radius: 6px;
+
+    background: #3f1d96;
+    color: white;
+    font-size: 14px;
+    text-align: center;
+    text-transform: capitalize;
+    font-weight: 600;
+
+    @media only screen and (max-width: 570px) {
+      width: 100%;
+    }
+
+    &:hover {
+      opacity: 0.85;
+    }
+
+    &:disabled {
+      color: #7c7b7c;
+      background: #ffffff;
+      border: 1px solid #7c7b7c;
+      cursor: not-allowed;
+
+      svg path {
+        fill: #7c7b7c;
+      }
+    }
+
+    .loading {
+      height: 18px;
+      width: 18px;
+    }
+
+    .buttonImageContainer {
+      margin-right: 5px;
+      height: 12px;
+      width: 12px;
+    }
+  }
+
+  .createTopicBody {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+
+    width: 100%;
+  }
+
+  .createTopicLabel {
+    padding-left: 4px;
+    padding-right: 4px;
+
+    font-size: 16px;
+    font-weight: 500;
+  }
+
+  .createTopicTitleInput {
+    width: 100%;
+    font-size: 14px;
+    padding: 12px;
+    margin-top: 4px;
+    margin-bottom: 16px;
+
+    border-color: #d1d5db;
+    border-width: 1px;
+    border-radius: 6px;
+  }
+
+  .createTopicTitleInput:focus {
+    outline: 2px solid transparent;
+    outline-offset: 2px;
+  }
+
+  .createTopicTextArea {
+    padding-top: 4px;
+    height: 144px;
+  }
+
+  .addTokenGateSelect {
+    padding: 8px;
+    font-size: 14px;
+    border-width: 1px;
+    border-color: #b8b9dc;
+    border-radius: 6px;
+    color: black;
+  }
+
+  .gateCheckbox {
+    display: flex;
+    align-items: center;
+    gap: 3px;
+    padding-bottom: 16px;
+
+    input {
+      margin-top: 2px;
+    }
+  }
+
+  .newAccessToken {
+    width: 100%;
+    padding: 12px;
+    margin: 4px 0 24px;
+
+    font-size: 14px;
+    border-width: 1px;
+    border-color: #d1d5db;
+    border-radius: 6px;
+
+    &:focus {
+      outline: 2px solid transparent;
+      outline-offset: 2px;
+    }
+  }
+
+  .okButton {
+    margin: 0 0 0 auto;
+    height: 48px;
+    padding: 0 16px;
+    border-radius: 6px;
+
+    background: #3f1d96;
+    color: white;
+    font-size: 14px;
+    text-align: center;
+    text-transform: capitalize;
+    font-weight: 600;
+
+    &:hover {
+      opacity: 0.85;
+    }
+
+    &:disabled {
+      cursor: not-allowed;
+      opacity: 0.6;
+    }
+
+    @media only screen and (max-width: 570px) {
+      width: 100%;
+    }
+  }
+
+  .cancelButton {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    height: 48px;
+    padding: 0 16px;
+    border-radius: 6px;
+    border-width: 1px;
+    border-color: #3f1d96;
+
+    background: white;
+    color: #3f1d96;
+    font-size: 14px;
+    font-weight: 600;
+    text-align: center;
+    text-transform: capitalize;
+
+    &:hover {
+      opacity: 0.85;
+    }
+
+    &:disabled {
+      cursor: not-allowed;
+      opacity: 0.6;
+    }
+
+    @media only screen and (max-width: 570px) {
+      width: 100%;
+    }
+  }
+}
