+++ src/styles/TopicList.css
@@ -88,6 +88,18 @@
     margin: auto;
   }
 
+  .inFlight {
+    opacity: 0.75;
+    height: 46px;
+    background: #bdddf445;
+    width: 100%;
+
+    text-align: left;
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+  }
+
   .emptyTopicList {
     width: 100%;
     min-width: 100%;
