+++ src/styles/customizations/DegenApeAcademy.css
@@ -242,6 +242,10 @@
       color: white;
     }
 
+    .row.topicInFlight .confirmingTopic {
+      color: white;
+    }
+
     .row {
       border-bottom: 1px solid white;
       font-weight: 400;
