+++ src/styles/customizations/Metaplex.css
@@ -240,6 +240,10 @@
       color: white;
     }
 
+    .row.topicInFlight .confirmingTopic {
+      color: white;
+    }
+
     .row {
       border-bottom: 1px solid white;
       font-weight: 400;
